requirements.txt
asgiref==3.5.2
black==22.12.0
click==8.1.3
colorama==0.4.6
Django==4.1.4
django-cors-headers==3.13.0
django-resized==1.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
pycodestyle==2.10.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
